Add unit tests and CI (#99)
* Add tests

* Rename tests to test

* Add tests

* Fix lint.yaml

* Add cpg-utils[workflows]

* Tests: moch exists and cohort; avoid writing files for dry_run

* Tests: use LocalBackend and mock cloudfuse (CI environment will not let setup ServiceBackend)

* Use standard tmpdir fixture

* Clean up

* Test: remove extra_config and set everything in DEFAULT_CONF. Use LocalBackend

* Test: clean up config

* Test: mock read_secret for es_password

+++ jobs/joint_genotyping.py
@@ -84,8 +84,9 @@ def make_joint_genotyping_jobs(
     df = pd.DataFrame(
         [{'id': sid, 'path': str(path)} for sid, path in gvcf_by_sid.items()]
     )
-    with sample_map_bucket_path.open('w') as fp:
-        df.to_csv(fp, index=False, header=False, sep='\t')
+    if not get_config()['hail'].get('dry_run', False):
+        with sample_map_bucket_path.open('w') as fp:
+            df.to_csv(fp, index=False, header=False, sep='\t')
 
     do_filter_excesshet = len(gvcf_by_sid) >= 1000 and do_filter_excesshet
 
@@ -223,7 +224,7 @@ def genomicsdb(
         job_name,
         (job_attrs or {})
         | dict(
-            tool='gatk_GenomicsDBImport',
+            tool='gatk GenomicsDBImport',
         ),
     )
     j.image(image_path('gatk'))
